separate ODE system and problem
separated ODE system and problem so that the system does not have to be built for each problem.
as a consequence variables and parameters are not defined globally instead of within the function.
removed obsolete tests and reorganized existing tests.
changed the inputs for total_payoff! and behav_eq!.
removed useless libraries; will add back as required.