small optimizations
collecting parameters also calculates the i, -i, and group norms at the same time so that `behav_eq` can use the collected values as is.
removed use of `vcat` and switched to manual unpacking and construction. Which results in fewer allocations, but slightly slower. Maybe the tradeoff is worth it for larger computations?
forced `objective_derivative` to be type stable.
Overall fewer allocations and memory are used, but only marginally.