fix: workaround for norm_i and norm_-i
norm for i and -i individuals is calculated incorrectly. It should be done on a per individual basis not per group basis. This increases the complexity and amount of data needed to be stored. Thus I have, for now, hard coded both values to be zero until I come up with a solution.
disabled mutation of internal punishment.