feat: add exporting snapshots
Add ability to save the simulation in states other than the end.
Allows multiple points to be selected including specific generations or percent completion.
if specified generation or percentage is referencing a generation not within the data then the data will not be exported since there is none and the preallocated dataframe will be deleted.